+++ src/Generators/Tests/Laravel/Policy/PolicyTestGenerator.php
@@ -6,10 +6,8 @@ namespace PhpUnitGen\Core\Generators\Tests\Laravel\Policy;
 
 use PhpUnitGen\Core\Aware\DocumentationFactoryAwareTrait;
 use PhpUnitGen\Core\Contracts\Aware\DocumentationFactoryAware;
-use PhpUnitGen\Core\Contracts\Generators\Factories\ClassFactory as ClassFactoryContract;
 use PhpUnitGen\Core\Contracts\Generators\Factories\MethodFactory as MethodFactoryContract;
-use PhpUnitGen\Core\Generators\Tests\Basic\BasicTestGenerator;
-use PhpUnitGen\Core\Generators\Tests\Laravel\UnitClassFactory;
+use PhpUnitGen\Core\Generators\Tests\Laravel\LaravelTestGenerator;
 use PhpUnitGen\Core\Generators\Tests\Laravel\UsesUserModel;
 use PhpUnitGen\Core\Models\TestClass;
 use PhpUnitGen\Core\Models\TestProperty;
@@ -22,7 +20,7 @@ use Roave\BetterReflection\Reflection\ReflectionMethod;
  * @author  Killian Hascoët <killianh@live.fr>
  * @license MIT
  */
-class PolicyTestGenerator extends BasicTestGenerator implements DocumentationFactoryAware
+class PolicyTestGenerator extends LaravelTestGenerator implements DocumentationFactoryAware
 {
     use DocumentationFactoryAwareTrait;
     use UsesUserModel;
@@ -33,7 +31,6 @@ class PolicyTestGenerator extends BasicTestGenerator implements DocumentationFac
     public static function implementations(): array
     {
         return array_merge(parent::implementations(), [
-            ClassFactoryContract::class  => UnitClassFactory::class,
             MethodFactoryContract::class => PolicyMethodFactory::class,
         ]);
     }